refact: ChatRoom 도메인의 Test API 분리 및 개발환경에서만 허용
