+++ tlog-was/src/main/java/com/se/Tlog/domain/Social/Chat/room/ChatRoomController.java
@@ -1,12 +1,5 @@
 package com.se.Tlog.domain.Social.Chat.room;
 
-import com.se.Tlog.domain.User.domain.Role;
-import com.se.Tlog.domain.User.domain.User;
-import com.se.Tlog.domain.User.repository.jpa.UserRepository;
-import com.se.Tlog.global.exception.CustomException;
-import com.se.Tlog.global.response.error.ErrorType;
-import com.se.Tlog.global.response.success.SuccessRes;
-import com.se.Tlog.global.util.jwt.AccessTokenProvider;
 import io.swagger.v3.oas.annotations.Operation;
 import io.swagger.v3.oas.annotations.Parameter;
 import io.swagger.v3.oas.annotations.responses.ApiResponse;
@@ -14,6 +7,8 @@ import io.swagger.v3.oas.annotations.security.SecurityRequirement;
 import lombok.RequiredArgsConstructor;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
+
+import com.se.Tlog.global.response.success.SuccessRes;
 
 import java.util.UUID;
 
@@ -23,8 +18,6 @@ import java.util.UUID;
 public class ChatRoomController {
 
     private final ChatRoomService chatRoomService;
-    private final AccessTokenProvider accessTokenProvider;
-    private final UserRepository userRepository;    // 테스트 때문에 넣어둠 나중에 지우겠습니다!
 
     @GetMapping("/room/{hostId}")
     @Operation(
@@ -43,17 +36,5 @@ public class ChatRoomController {
     )
     public ResponseEntity<?> getRoomList(@PathVariable(name = "hostId") UUID hostId) {
         return ResponseEntity.ok().body(SuccessRes.from(chatRoomService.getRoomList(hostId)));
-    }
-    // test 편의성 때문에 토큰 생성 api 작성
-    @PostMapping("/{userId}")
-    public ResponseEntity<?> createToken(@PathVariable(name = "userId") UUID userId) {
-        User user = userRepository.findById(userId).orElseThrow(() -> new CustomException(ErrorType.NOT_FOUND));
-        String snsId = user.getSnsId();
-        return ResponseEntity.ok()
-                .body(accessTokenProvider.generateToken(
-                        userId.toString(),
-                        Role.USER.getValue(),
-                        snsId == null ? "" : snsId
-                ));
     }
 }
